# control-plane/api/openapi.yaml
openapi: 3.0.3
info:
  title: Peristera Control Plane API
  description: >
    Tenant lifecycle management. Everything the control-plane UI can do is
    here (API-first, README §4); versioning rules per ADR-0007 — additive
    within v1, breaks logged in the worklog until the first external
    consumer (M6). The tenant slug is the permanent identifier (ADR-0007
    §4), so it is the path parameter.
  version: "1.0"
servers:
  - url: /api/v1
security:
  - bearer: []
paths:
  /tenants:
    get:
      operationId: listTenants
      summary: List all tenants
      responses:
        "200":
          description: All tenants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createTenant
      summary: Create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantCreate"
      responses:
        "201":
          description: Created; provisioning continues asynchronously.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: A tenant with this slug already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tenants/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$"
    get:
      operationId: getTenant
      summary: Get one tenant
      responses:
        "200":
          description: The tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteTenant
      summary: Off-board a tenant (removes all its data)
      responses:
        "204":
          description: Deletion started; teardown continues asynchronously.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tenants/{slug}/users:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$"
    post:
      operationId: createTenantUser
      summary: Create an admin user in a tenant
      description: >
        Creates a human user (login = email) as an ORG_OWNER in the tenant's
        own Zitadel instance and returns a generated one-time password — the
        operator's handover artifact. Use it for a tenant's first admin and for
        lost-login recovery. The password is returned ONCE and never stored; if
        lost, create another user. Requires the tenant to be provisioned (it
        must have an issuer).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUserCreate"
      responses:
        "201":
          description: User created; credentials returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUserCredentials"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: A user with this email already exists in the tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: The tenant is not provisioned yet (no issuer).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        A Zitadel access token (or PAT) of an operator in the default
        instance. Browser sessions authenticate with the session cookie
        instead.
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such tenant.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    TenantCreate:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$"
          description: Permanent identifier; forms the tenant domain (immutable).
        displayName:
          type: string
        domain:
          type: string
          pattern: '^([a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?\.)+[a-z]{2,}$'
          description: >
            Optional custom apex (BYO domain, e.g. peristera.lu). When set, the
            tenant's issuer and apps live under it (<app>.<domain>) instead of
            the default <slug>.<platform-base>. Immutable. The domain must
            resolve to the platform so certs can issue.
    Tenant:
      type: object
      required: [slug, phase, permalink]
      properties:
        slug: { type: string }
        displayName: { type: string }
        phase:
          type: string
          enum: [Pending, Ready, Deleting, Failed, ""]
        issuer:
          type: string
          description: The tenant's OIDC issuer URL (permanent).
        clientId: { type: string }
        permalink:
          type: string
          description: Canonical URL of this resource (ADR-0007).
    TenantUserCreate:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Login and contact email of the user to create.
    TenantUserCredentials:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          description: The user's login name (their email).
        password:
          type: string
          description: >
            Generated one-time password (change required at first login).
            Returned once and never stored — relay it to the tenant.
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }